feat(admin): implement user management and enhance admin capabilities
This commit introduces a new administrative user management system via a
Supabase Edge Function (`admin-users`) allowing admins to create and
manage auth users directly. It also expands the admin suite with
improved RAG (Retrieval-Augmented Generation) management and news
editing tools.

Key changes:
- feat(admin): add `admin-users` edge function for CRUD operations on
  auth users.
- feat(admin): implement `AdminUsers.vue` for managing user profiles
  and roles.
- feat(ai): upgrade LLM model to `llama-3.3-70b-versatile` and improve
  rate-limit handling in `chat` edge function.
- feat(ai): add explicit rate-limit error messaging in `AiAssistant.vue`.
- feat(news): enhance news editing with a more robust `RichTextEditor`
  and improved sanitization.
- refactor(reactions): optimize `useReactions.js` to prevent stale
  counts by recomputing totals from scratch.
- refactor(ui): clean up code comments, documentation, and styling
  across the project for better maintainability.
- fix(db): update schema and migrations to support enhanced RAG
  functionality and like/dislike types.
- chore: add `demo_video_script.md` and remove obsolete `AdminSettings.vue`.

+++ supabase/rag_setup.sql
@@ -1,5 +1,5 @@
 -- =============================================================================
--- RAG Production Setup — run in Supabase Studio > SQL Editor
+-- RAG setup — run in Supabase Studio > SQL Editor
 -- =============================================================================
 
 -- 0) Unique constraint for upsert (idempotent)
@@ -39,12 +39,12 @@ create or replace function match_documents_fts(
   query_text text,
   match_count int default 20
 ) returns table (
-  id bigint, source text, source_id text, title text, content text, rank float
+  id bigint, source text, source_id text, title text, content text, rank float8
 ) language plpgsql as $$
 begin
   return query
   select d.id, d.source, d.source_id, d.title, d.content,
-    ts_rank(to_tsvector('english', coalesce(d.content, '')), plainto_tsquery('english', query_text)) as rank
+    ts_rank(to_tsvector('english', coalesce(d.content, '')), plainto_tsquery('english', query_text))::float8 as rank
   from documents d
   where d.embedding is not null
     and to_tsvector('english', coalesce(d.content, '')) @@ plainto_tsquery('english', query_text)
